Fix signedness warning in gio/tests/file.c
gio/tests/file.c: In function ‘written_cb’:
gio/tests/file.c:358:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘size_t’ {aka ‘long unsigned int’}
  358 |   if (data->pos < strlen (data->data))
      |                 ^

+++ gio/tests/file.c
@@ -174,7 +174,7 @@ typedef struct
   gint monitor_deleted;
   gint monitor_changed;
   gchar *monitor_path;
-  gint pos;
+  gsize pos;
   const gchar *data;
   gchar *buffer;
   guint timeout;
